databricks-functions-dataexplorer/tools/fakedata-generator/requirements.txt
azure-storage-blob==12.6.0
apscheduler==3.6.3